int -> long. Not going to happen of course.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1823283 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/tomcat/util/net/Nio2Endpoint.java
@@ -878,7 +878,7 @@ public class Nio2Endpoint extends AbstractJsseEndpoint<Nio2Channel,AsynchronousS
         private class ScatterReadCompletionHandler<A> implements CompletionHandler<Long, OperationState<A>> {
             @Override
             public void completed(Long nBytes, OperationState<A> state) {
-                if (nBytes.intValue() < 0) {
+                if (nBytes.longValue() < 0) {
                     failed(new EOFException(), state);
                 } else {
                     state.nBytes += nBytes.longValue();
